fix: move type guards outside logger
